\name{tclust.test.data}
\alias{tclust.test.data}
\docType{data}
\title{test data for tight clustering package}
\description{
Sample microarray data
}
\usage{data(tclust.test.data)}
\format{
The data is a list of 3 items:
  \describe{
    \item{GeneID}{ID of each gene} 
    \item{Annotation}{Annotation information of each gene} 
    \item{Data}{Data matirx of gene expression: each row represent one gene; each column represent one sample} 
  }
}

\keyword{datasets}
