\name{tightClust-package}
\alias{tightClust-package}
\alias{tightClust}
\docType{package}
\title{
Tight Clustering Package
}
\description{
This package could perform tight clustering algorithm proposed by George C. Tseng and Wing H. Wong.
}
\details{
\tabular{ll}{
Package: \tab tightClust\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-28\cr
License: \tab GPL (>=2)\cr
}
}
\author{
George C. Tseng <ctseng@pitt.edu>, Wing H. Wong <whwong@stanford.edu>

Maintainer: Chi Song <chs108@pitt.edu>
}
\references{
George C. Tseng and Wing H. Wong. (2005) Tight Clustering: A Resampling-based Approach for Identifying Stable and Tight Patterns in Data.  Biometrics.61:10-16.
}
\keyword{ package }
