\name{write.tight.clust}
\alias{write.tight.clust}

\title{Write tight cluster result}
\description{
  A function to print the tight cluster result to a file or connection.
}
\usage{
write.tight.clust(x, ...)
}

\arguments{
  \item{x}{Return value of the \code{tight.clust} function}
  \item{\dots}{Arguments to \code{write.table}.}
}
\references{George C. Tseng and Wing H. Wong. (2005) Tight Clustering: A Resampling-based Approach for Identifying Stable and Tight Patterns in Data.  Biometrics.61:10-16.}
\author{Chi Song <chs108@pitt.edu>}

\keyword{ cluster	}